Add test for map-style samples reheader
`bcftools reheader` `--samples` does not only allow to to provide
a file with the new names (assuming the names in the file match the
header), but also allow a file with a whitespace separated map:
```
<old name> <new_name>\n
...
```
enabling a more expliclit renaming.

This used to be supported by `snakemake-wrappers` as well (eg used
in workflow: https://github.com/nikostr/dna-seq-deepvariant-glnexus-variant-calling).